+++ CHANGELOG.md
@@ -6,9 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.11.1] - 10.07.2021
 ### Changed
-- Tried to fix edge cases where unhandled exception was thrown ([Related to Node-RED project, issue #13](https://github.com/jisotalo/node-red-contrib-ads-client/issues/13))
+- Tried to fix edge cases where unhandled exception was thrown ([node-red-contrib-ads-client, issue #13](https://github.com/jisotalo/node-red-contrib-ads-client/issues/13))
   - Example: When PLC was restarted (config mode) multiple times, sometimes an exception was thrown
 - Improved operation during connection faults
+- Fixed typos ([See issue #71](https://github.com/jisotalo/ads-client/issues/71))
+  - `byteArrayToAmsNedIdStr` --> `byteArrayToAmsNetIdStr`
+  - `amsNedIdStrToByteArray `--> `amsNetIdStrToByteArray`
 
 
 ## [1.11.0] - 29.06.2021